refactor: fix classnames and method names
Follow auto routing naming conventions.
